+++ packages/blocks/feature-section-grid/src/api-harmonization/index.ts
@@ -1,0 +1,8 @@
+export const URL = '/blocks/feature-section-grid';
+
+export { FeatureSectionGridBlockModule as Module } from './feature-section-grid.module';
+export { FeatureSectionGridService as Service } from './feature-section-grid.service';
+export { FeatureSectionGridController as Controller } from './feature-section-grid.controller';
+
+export * as Model from './feature-section-grid.model';
+export * as Request from './feature-section-grid.request';