+++ packages/blocks/notification-summary/src/api-harmonization/index.ts
@@ -1,0 +1,8 @@
+export const URL = '/blocks/notification-summary';
+
+export { NotificationSummaryBlockModule as Module } from './notification-summary.module';
+export { NotificationSummaryService as Service } from './notification-summary.service';
+export { NotificationSummaryController as Controller } from './notification-summary.controller';
+
+export * as Model from './notification-summary.model';
+export * as Request from './notification-summary.request';