+++ packages/blocks/surveyjs-form/lint-staged.config.mjs
@@ -1,4 +1,3 @@
-export default {
-    '*.{js,jsx,ts,tsx,css,scss}': ['prettier --write'],
-    '*.{js,jsx,ts,tsx}': () => 'tsc --noEmit',
-};
+import { config } from '@o2s/lint-staged-config/base';
+
+export default config;