+++ packages/ui/src/components/Cards/PricingCard/PricingCard.tsx
@@ -49,8 +49,20 @@ export const FeatureItem: React.FC<FeatureItemProps> = ({ title, description, ic
 };
 
 export const PricingCard: React.FC<Readonly<PricingCardProps>> = (props) => {
-    const { title, description, image, price, isPromoted, tags, links, featureListTitle, featureList, LinkComponent } =
-        props;
+    const {
+        title,
+        titleType: TitleTag = 'h3',
+        description,
+        image,
+        price,
+        isPromoted,
+        tags,
+        links,
+        featureListTitle,
+        featureList,
+        LinkComponent,
+        hasPriority,
+    } = props;
 
     return (
         <div
@@ -68,6 +80,7 @@ export const PricingCard: React.FC<Readonly<PricingCardProps>> = (props) => {
                             width={image.width}
                             height={image.height}
                             className="object-cover object-center w-full h-full"
+                            priority={hasPriority}
                         />
                     </div>
                 )}
@@ -76,7 +89,7 @@ export const PricingCard: React.FC<Readonly<PricingCardProps>> = (props) => {
                     <div className="flex flex-col gap-3">
                         {title && (
                             <Typography variant="highlightedSmall" asChild className={cn(isPromoted && 'text-primary')}>
-                                <h3>{title}</h3>
+                                <TitleTag>{title}</TitleTag>
                             </Typography>
                         )}