+++ packages/cli/create-o2s-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # create-o2s-app
 
+## 1.1.4
+
+### Patch Changes
+
+- 221dc2c: fix: added missing Eslint command to list-staged
+- Updated dependencies [221dc2c]
+    - @o2s/telemetry@1.1.2
+
+## 1.1.3
+
+### Patch Changes
+
+- 0354126: reworked turbo dev task to optimize CPU/RAM usage
+- Updated dependencies [0354126]
+    - @o2s/telemetry@1.1.1
+
 ## 1.1.2
 
 ### Patch Changes